backend/requirements.txt
django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
drf-nested-routers==0.95.0
psycopg2-binary==2.9.11
Pillow==10.3.0
python-decouple==3.8
